+++ src/SecucardConnect/Product/Payment/ContractsService.php
@@ -6,10 +6,12 @@ use GuzzleHttp\Exception\GuzzleException;
 use SecucardConnect\Client\ApiError;
 use SecucardConnect\Client\AuthError;
 use SecucardConnect\Client\ClientError;
+use SecucardConnect\Client\MissingParamsError;
 use SecucardConnect\Client\ProductService;
 use SecucardConnect\Product\Payment\Model\CloneParams;
 use SecucardConnect\Product\Payment\Model\CreateSubContractRequest;
 use SecucardConnect\Product\Payment\Model\CreateSubContractResponse;
+use SecucardConnect\Product\Payment\Model\Transaction;
 use SecucardConnect\Product\Services\Model\Contract;
 
 
@@ -83,5 +85,27 @@ class ContractsService extends ProductService
         }
 
         return $this->execute($contract_id, 'requestId', null, $param, CreateSubContractResponse::class);
+    }
+
+    /**
+     * Remove the accrual flag from all payment transactions of the given contract
+     * (this will be done in the background,
+     *  so the response TRUE will only indicate the the API has accepted this request)
+     *
+     * @param string $contractId The payment contract id
+     * @return bool
+     * @throws ApiError
+     * @throws AuthError
+     * @throws ClientError
+     * @throws GuzzleException
+     * @throws MissingParamsError
+     */
+    public function revokeAccrual($contractId)
+    {
+        if (empty($contractId)) {
+            throw new MissingParamsError('contractId', __METHOD__);
+        }
+
+        return (bool)$this->execute($contractId, 'revokeAccrual');
     }
 }