+++ CHANGELOG.md
@@ -18,6 +18,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Removed
 
+## [1.22.0] - 2021-06-11
+
+### Added
+- General.ContractsService: new method `revokeAccrual()`
+- Payment.ContractsService: new method `revokeAccrual()`
+
+### Changed
+- Adjusted response handling to support empty HTTP responses like `204 No Content`
+
 
 ## [1.21.0] - 2021-03-10
 
@@ -693,3 +702,4 @@ First release
 [1.19.0]:https://github.com/secucard/secucard-connect-php-sdk/compare/1.18.0...1.19.0
 [1.20.0]:https://github.com/secucard/secucard-connect-php-sdk/compare/1.19.0...1.20.0
 [1.21.0]:https://github.com/secucard/secucard-connect-php-sdk/compare/1.20.0...1.21.0
+[1.22.0]:https://github.com/secucard/secucard-connect-php-sdk/compare/1.21.0...1.22.0